Navigator 폴더 삭제 / App.js에 통합
