+++ App.js
@@ -1,6 +1,31 @@
-import React from 'react';
-import AppNavigator from './src/navigation/AppNavigator';
+import { StyleSheet, Text, View } from "react-native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { NavigationContainer } from "@react-navigation/native";
+import MainPage from "./screen/MainPage";
+import SearchPage from "./screen/SearchPage";
+import Footer from "./components/Footer";
+import LoadingPage from './screen/LoadingPage';
+
+const Stack = createNativeStackNavigator();
 
 export default function App() {
-  return <AppNavigator />;
+  //npm install @react-navigation/native
+  //npm install @react-navigation/native-stack
+  return (
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="LoadingPage" screenOptions={{ headerShown: false }}>
+        <Stack.Screen name="LoadingPage" component={LoadingPage} />
+        <Stack.Screen name="MainPage" component={MainPage} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+});
