--- src/components/FishInfo.js
@@ -1,11 +1,0 @@
-import { View, Text } from "react-native";
-
-export default function FishInfo() {
-  //물고기 정보
-  return (
-    //출현 장소 / 출현 시기 / 출현 시간대 / 물고기 이름
-    <View>
-      <Text></Text>
-    </View>
-  );
-}
